Add files via upload
This is one of my coolest/best projects yet with HTML. I created a home page with a profile picture and on it you are able to view a resume page and a projects page which will show you two different projects I have worked on. I also used a bootstrap template to help me create this.